replace sync with a time based incremental one
